+++ src/viewers/VivViewer.js
@@ -35,6 +35,7 @@ export default class VivViewer extends PureComponent {
     });
     this._onViewStateChange = this._onViewStateChange.bind(this);
     this.layerFilter = this.layerFilter.bind(this);
+    this.onHover = this.onHover.bind(this);
   }
 
   /**
@@ -107,14 +108,60 @@ export default class VivViewer extends PureComponent {
     return prevState;
   }
 
+  // eslint-disable-next-line consistent-return
+  onHover({ sourceLayer, coordinate, layer }) {
+    if (!coordinate) {
+      return null;
+    }
+    const { hoverHooks } = this.props;
+    if (!hoverHooks) {
+      return null;
+    }
+    const { handleValue } = hoverHooks;
+    if (!handleValue) {
+      return null;
+    }
+    const { channelData, bounds } = sourceLayer.props;
+    const { data, width } = channelData;
+    let dataCoords;
+    // This is currently a work-around for: https://github.com/visgl/deck.gl/pull/4526.
+    // Once this is fixed we can make something more robust.
+    if (sourceLayer.id.includes('Background')) {
+      const { numLevels } = layer.props.loader;
+      // The zoomed out layer needs to use the fixed zoom at which it is rendered (i.e numLevels - 1).
+      const layerZoomScale = Math.max(1, 2 ** Math.floor(numLevels - 1));
+      dataCoords = [
+        Math.floor((coordinate[0] - bounds[0]) / layerZoomScale),
+        Math.floor((coordinate[1] - bounds[3]) / layerZoomScale)
+      ];
+    } else {
+      // Using floor means that as we zoom out, we are scaling by the zoom just passed, not the one coming.
+      const { zoom } = layer.context.viewport;
+      const layerZoomScale = Math.max(1, 2 ** Math.floor(-zoom));
+      dataCoords = [
+        Math.floor((coordinate[0] - bounds[0]) / layerZoomScale),
+        Math.floor((coordinate[1] - bounds[3]) / layerZoomScale)
+      ];
+    }
+    const hoverData = data.map(d => d[dataCoords[1] * width + dataCoords[0]]);
+    handleValue(hoverData);
+  }
+
   /**
    * This renders the layers in the DeckGL context.
    */
   _renderLayers() {
+    const { onHover } = this;
     const { viewStates } = this.state;
     const { views, layerProps } = this.props;
     return views.map((view, i) =>
-      view.getLayers({ viewStates, props: layerProps[i] })
+      view.getLayers({
+        viewStates,
+        props: {
+          ...layerProps[i],
+          onHover
+        }
+      })
     );
   }
 
@@ -148,6 +195,9 @@ export default class VivViewer extends PureComponent {
         onViewStateChange={this._onViewStateChange}
         views={deckGLViews}
         viewState={viewStates}
+        getCursor={({ isDragging }) => {
+          return isDragging ? 'grabbing' : 'crosshair';
+        }}
       />
     ) : (
       <div className="viv-error">